ACP-L3-FS-SEAL-TARGET-EXECUTION-PORT: split ondemandtarget test file below backend-size gate
Round-6 review flagged service_test.go growing to 1,079 lines, crossing
the repository's 1,000-line backend-size limit. Move the invocation,
cancellation, close, and published-capability tests into a new
target_execution_test.go, leaving the New/StartAsync tests and shared
fakes/helpers in service_test.go. Both files now sit at ~570/~525 lines.
